+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-06-10 09:58:06
-// Commit: docs(ui): update readme and API contracts for ui
+// Updated at: 2026-06-10 12:26:25
+// Commit: feat(ui): add bi-directional event bus handler for ui
 
 class uiComponentState {
-  final String commitRef = '1863';
-  final String lastMessage = 'docs(ui): update readme and API contracts for ui';
-  final DateTime timestamp = DateTime.parse('2026-06-10T09:58:06+05:00');
+  final String commitRef = '1868';
+  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
+  final DateTime timestamp = DateTime.parse('2026-06-10T12:26:25+05:00');
 }